requirements.txt
alembic==1.17.2
fastapi==0.123.0
instructor>=1.13.0
openai==2.9.0
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.0
slowapi==0.1.9
SQLAlchemy==2.0.43
uvicorn==0.38.0
